https://ticketbookin.ccbp.tech/
We expect the developer to create a ticket booking platform that caters to both event
organizers and audience members. The core of this project is to efficiently manage
concurrent ticket bookings while providing a fast and robust notification system.
Candidates are encouraged to use technologies such as Redis, Kafka, or any other suitable
services to achieve high performance and scalability. The frontend must be developed with
React, reflecting a strong preference for a MERN stack, though flexibility in the choice of
backend technologies is allowed. The focus is on functionality and performance, with UI
design being a secondary concern.